+++ extensions/mattermost/CHANGELOG.md
@@ -6,7 +6,7 @@
 - `signIn` function is now `async`
 - Show better error message when Authorization fails (should close [#14876](https://github.com/raycast/extensions/issues/14876))
 - `usePromise` in `withAuthorization` component for better control
-- In `search-channels` you can specify in Preferences whether to launch installed Mattermost application or not
+- You can now specify in Preferences whether to open deep links in Browser or Mattermost Application
 - Fix: various typos
 - Add an `Axios` intercept for robust error handling
 - Update dependencies